Fix promotion issues in iOS unit tests

+++ Tests/BKAsyncTestCase.m
@@ -102,7 +102,7 @@ typedef NS_ENUM(NSUInteger, SenTestCaseError) {
 	if (error == SenTestCaseErrorTimedOut) {
 		XCTFail(@"Request timed out");
 	} else if (error == SenTestCaseErrorInvalidStatus) {
-		XCTFail(@"Request finished with the wrong status: %ld != %ld", (long)status, notifiedStatus_);
+		XCTFail(@"Request finished with the wrong status: %ld != %ld", (long)status, (long)notifiedStatus_);
 	} else if (error == SenTestCaseErrorUnprepared) {
 		XCTFail(@"Call prepare before calling asynchronous method and waitForStatus:timeout:");
 	}
